pyproject.toml
[project]
name = "memepool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.4",
    "peewee>=3.17.9",
    "praw>=7.8.1",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.5",
    "tweepy>=4.15.0",
    "twikit>=2.3.3",
    "web3>=7.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.5",
]
